requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.5
Django==3.1.6
django-extensions==3.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
